Add files via upload
This file should be run along singleexpressionconversion. It accepts a command line input as follows : python3 fullexpressionparsing.py t2.txt, where t2.txt contains our latex code which is to be converted into a python code. This creates a file named GeneratedCodefort2.py, which contains it's corresponding einsum.
One area where this could be generalized is that it can only deal up to 2 permutations of 2 indices each. This has to be generalized to any number of permutations with any number of indices each. One reference for this code is : https://note.nkmk.me/en/python-split-rsplit-splitlines-re/